New bloom filter

+++ processor/structs.go
@@ -45,11 +45,11 @@ type LanguageFeature struct {
 	Strings               *Trie
 	Tokens                *Trie
 	Nested                bool
-	ComplexityCheckMask   byte
-	SingleLineCommentMask byte
-	MultiLineCommentMask  byte
-	StringCheckMask       byte
-	ProcessMask           byte
+	ComplexityCheckMask   uint64
+	SingleLineCommentMask uint64
+	MultiLineCommentMask  uint64
+	StringCheckMask       uint64
+	ProcessMask           uint64
 	Keywords              []string
 	Quotes                []Quote
 }
